requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi
uvicorn

sentence-transformers
torch
transformers
huggingface-hub

faiss-cpu
numpy
pandas
pyarrow

requests
python-dotenv
lxml
beautifulsoup4

langchain
langchain-openai
langchain-core
langgraph

pydantic
tiktoken
rank-bm25
redis
qdrant-client
ragas
datasets
langchain
openai
